refactor rate limiter
